feat: refactorizar gestión de modales y simplificar código en main.ts
🔧 Se han eliminado funciones obsoletas relacionadas con la gestión de modales y se ha implementado un sistema de montaje y limpieza para el modal de configuración. Además, se ha optimizado la creación de componentes de UI, mejorando la eficiencia y la claridad del código. Estos cambios contribuyen a una mejor experiencia de usuario al interactuar con la interfaz de configuración.
